% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HotellingEllipse.R
\name{HotellingEllipse}
\alias{HotellingEllipse}
\title{Hotelling Ellipse}
\usage{
HotellingEllipse(data, k = 2, pcx = 1, pcy = 2)
}
\arguments{
\item{data}{Data frame or tibble of PCA/PLS scores coordinates.}

\item{k}{Integer, specifying the number of components (k = 2 by default).}

\item{pcx}{Integer, specifying which component is on the x-axis (pcx = 1 by default).}

\item{pcy}{Integer, specifying which component is on the y-axis (pcy = 2 by default).}
}
\value{
A list with the T-squared statistic, Hotelling ellipse axis parameters, and T-squared cutoff at 95\% and 99\% confidence levels.
}
\description{
Computes axis parameters of Hotelling ellipse for a bivariate scatter plot.
The function is created to be used with PCA or PLS scores coordinates, i.e. on the PCs or LVs axes.
It shows the statistical limits computed using Hotelling T-squared distribution at 95\% and 99\% confidence levels.
}
\examples{
HotellingEllipse(data = PCA_scores, k = 2, pcx = 1, pcy = 2)

}
