% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HotellingEllipseCoord.R
\name{HotellingEllipseCoord}
\alias{HotellingEllipseCoord}
\title{T-squared Confidence Ellipse Coordinates}
\usage{
HotellingEllipseCoord(data, pcx = 1, pcy = 2, conf.limit = 0.95, pts = 200)
}
\arguments{
\item{data}{a data frame or tibble of scores coordinates}

\item{pcx}{an integer specifying which component is on the x-axis (by default 1)}

\item{pcy}{an integer specifying which component is on the y-axis (by default 2)}

\item{conf.limit}{an integer specifying the confidence level (by default 0.95)}

\item{pts}{an integer indicating the number of points in the ellipse (by default 200)}
}
\value{
a data frame with x and y coordinates
}
\description{
This function produces coordinates values for plotting confidence ellipse.
}
\details{
Coordinates for Hotelling's confidence ellipse
}
\examples{
#' ## Principal components analysis (PCA)
library(magrittr)
set.seed(123)
pca_mod <- LIBS_spec \%>\%
  dplyr::select(where(is.numeric)) \%>\%
  FactoMineR::PCA(scale.unit = FALSE, graph = FALSE)

## Extract PCA scores
pca_scores <- pca_mod \%>\%
   purrr::pluck("ind", "coord") \%>\%
   tibble::as_tibble()
## Compute Hotelling T-squared statistic and ellipse parameters
library(HotellingEllipse)
xy_coord <- HotellingEllipseCoord(data = pca_scores, pcx = 1, pcy = 2, conf.limit = 0.95, pts = 200)

}
\author{
Christian L. Goueguel,
christian.goueguel@gmail.com
}
