% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIBS_spec_data.R
\docType{data}
\name{LIBS_spec}
\alias{LIBS_spec}
\title{Laser-induced breakdown spectroscopy (LIBS) spectra of soils}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 171 rows and 6669 columns.
}
\usage{
data(LIBS_spec)
}
\description{
Data from a LIBS experiment, with 171 soil samples recorded at 240-820 nm wavelength range.
}
\examples{
data(LIBS_spec)
}
\keyword{datasets}
