% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIBS_spec.R
\docType{data}
\name{specData}
\alias{specData}
\title{LIBS spectra of 171 soil samples.}
\format{
A data frame with 171 rows and 6669 variables:
\describe{
\item{rows}{soil samples}
\item{columns}{wavelengths, nm}
}
}
\source{
\url{https://doi.org/10.1039/C9JA00090A}
}
\usage{
specData
}
\description{
A dataset containing the LIBS spectra of 171 agricultural soils.
}
\keyword{datasets}
