% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HotellingEllipse-package.R
\docType{package}
\name{HotellingEllipse-package}
\alias{HotellingEllipse}
\alias{HotellingEllipse-package}
\title{HotellingEllipse: Hotelling's T-Squared Statistic and Ellipse}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to calculate the Hotelling's T-squared statistic and corresponding 2D or 3D confidence ellipses. Provides the semi-axes of Hotelling's T-squared ellipse at 95% and 99% confidence levels. Additionally, enables users to obtain the coordinates in two or three dimensions at user-defined confidence levels, allowing for the construction of 2D or 3D ellipses with customized confidence levels. Bro and Smilde (2014) \doi{10.1039/c3ay41907j}. Brereton (2016) \doi{10.1002/cem.2763}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChristianGoueguel/HotellingEllipse}
  \item Report bugs at \url{https://github.com/ChristianGoueguel/HotellingEllipse/issues}
}

}
\author{
\strong{Maintainer}: Christian L. Goueguel \email{christian.goueguel@gmail.com} (\href{https://orcid.org/0000-0003-0521-3446}{ORCID})

}
\keyword{internal}
