% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HotellingEllipse.R
\name{HotellingEllipse}
\alias{HotellingEllipse}
\title{Hotelling's T-squared Ellipse}
\usage{
HotellingEllipse(data, k = 2, pcx = 1, pcy = 2)
}
\arguments{
\item{data}{a data frame or tibble of scores coordinates}

\item{k}{number of components (by default 2)}

\item{pcx}{an integer specifying which component is on the x-axis (by default 1)}

\item{pcy}{an integer specifying which component is on the y-axis (by default 2)}
}
\value{
Returns a list including:
(1) Tsquared, a data frame containing the T-squared statistic.
(2) Ellipse, a data frame containing the major and minor semi-axis values for both Hotelling T-squared ellipses.
(3) cutoff.99pct, an integer indicating the T-squared cutoff value at 99\% confidence level.
(4) cutoff.95pct, an integer indicating the T-squared cutoff value at 95\% confidence level.
}
\description{
Computes axis parameters of Hotelling ellipse for a bivariate scatter plot.
The function is created to be used with principal components analysis (PCA) or partial least squares (PLS) scores coordinates.
It shows the statistical limits computed using Hotelling T-squared distribution at 95\% and 99\% confidence levels.
}
\details{
Hotelling's T-squared Statistic and Confidence Ellipse
}
\examples{
## Principal components analysis (PCA)
library(magrittr)
set.seed(123)
pca_mod <- LIBS_spec \%>\%
  dplyr::select(where(is.numeric)) \%>\%
  FactoMineR::PCA(scale.unit = FALSE, graph = FALSE)

## Extract PCA scores
pca_scores <- pca_mod \%>\%
   purrr::pluck("ind", "coord") \%>\%
   tibble::as_tibble()

## Compute Hotelling T-squared statistic and ellipse parameters
library(HotellingEllipse)
T2 <- HotellingEllipse(data = pca_scores, k = 2, pcx = 1, pcy = 2)
}
\author{
Christian L. Goueguel,
christian.goueguel@gmail.com
}
