% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specData}
\alias{specData}
\title{LIBS spectra of 100 soil samples}
\format{
Data frame of 100 rows (soil samples) and 3152 columns (wavelengths).
}
\source{
\doi{10.1039/C9JA00090A}
}
\usage{
specData
}
\description{
A dataset containing the emission spectra of 171 soils measured in laboratory conditions.
The samples were cleaned, dried, homogenized, sieved (10 Mesh size) and thereafter pelletized prior to LIBS measurements.
LIBS spectra were preprocessed by performing baseline removal.
}
\keyword{datasets}
